latex2pdf/requirements.txt
django<3

# ImageField
Pillow

# Django and MongoDB database connector 
djongo==1.2.31
pymongo==3.7.2
sqlparse==0.2.4  # this is required by django 1.3.2

# For ImageMagick
wand

# For mypy (static type checking) support
typing>=3.6.1

# django forms
django-crispy-forms

# REST api
djangorestframework

# CORS
django-cors-headers

# codemirror
django-codemirror-widget>=0.5

# For web dependencies
django-npm

# For cache
python-memcached
django-redis

# For web storage
django-s3-storage
